+++ src/config/diff_ignore_whitespace_setting.rs
@@ -1,0 +1,6 @@
+#[derive(Clone, PartialEq, Debug)]
+pub(crate) enum DiffIgnoreWhitespaceSetting {
+	None,
+	All,
+	Change,
+}