+++ src/show_commit/mod.rs
@@ -1,10 +1,16 @@
 mod commit;
+mod delta;
+mod diff_line;
 mod file_stat;
+mod file_stats_builder;
+mod show_commit_state;
 mod status;
 mod user;
 mod util;
 mod view_builder;
 
+use crate::config::diff_ignore_whitespace_setting::DiffIgnoreWhitespaceSetting;
+use crate::config::diff_show_whitespace_setting::DiffShowWhitespaceSetting;
 use crate::config::Config;
 use crate::constants::MINIMUM_FULL_WINDOW_WIDTH;
 use crate::display::display_color::DisplayColor;
@@ -16,7 +22,8 @@ use crate::process::process_module::ProcessModule;
 use crate::process::process_result::{ProcessResult, ProcessResultBuilder};
 use crate::process::state::State;
 use crate::show_commit::commit::{Commit, LoadCommitDiffOptions};
-use crate::show_commit::view_builder::ViewBuilder;
+use crate::show_commit::show_commit_state::ShowCommitState;
+use crate::show_commit::view_builder::{ViewBuilder, ViewBuilderOptions};
 use crate::view::line_segment::LineSegment;
 use crate::view::view_data::ViewData;
 use crate::view::view_line::ViewLine;
@@ -28,6 +35,7 @@ pub(crate) struct ShowCommit<'s> {
 	view_data: ViewData,
 	no_commit_view_data: ViewData,
 	view_builder: ViewBuilder<'s>,
+	state: ShowCommitState,
 }
 
 impl<'s> ProcessModule for ShowCommit<'s> {
@@ -45,11 +53,16 @@ impl<'s> ProcessModule for ShowCommit<'s> {
 		self.commit = Some(Commit::new_from_hash(
 			git_interactive.get_selected_line_hash().as_str(),
 			LoadCommitDiffOptions {
-				renames: self.config.git.diff_renames,
-				rename_limit: self.config.git.diff_rename_limit,
+				context_lines: self.config.git.diff_context,
 				copies: self.config.git.diff_copies,
+				ignore_whitespace: self.config.diff_ignore_whitespace == DiffIgnoreWhitespaceSetting::All,
+				ignore_whitespace_change: self.config.diff_ignore_whitespace == DiffIgnoreWhitespaceSetting::Change,
+				interhunk_lines: self.config.git.diff_interhunk_lines,
+				rename_limit: self.config.git.diff_rename_limit,
+				renames: self.config.git.diff_renames,
 			},
 		));
+		self.state = ShowCommitState::Overview;
 	}
 
 	fn process(&mut self, _git_interactive: &mut GitInteractive, _: &View) -> ProcessResult {
@@ -77,12 +90,25 @@ impl<'s> ProcessModule for ShowCommit<'s> {
 			Input::Help => {
 				result = result.help(State::ShowCommit);
 			},
+			Input::ShowDiff => {
+				self.view_data.reset();
+				self.state = match self.state {
+					ShowCommitState::Overview => ShowCommitState::Diff,
+					ShowCommitState::Diff => ShowCommitState::Overview,
+				}
+			},
 			Input::Resize => {
 				self.view_data.clear();
 			},
 			_ => {
-				self.view_data.clear();
-				result = result.state(State::List(false));
+				if self.state == ShowCommitState::Diff {
+					self.view_data.reset();
+					self.state = ShowCommitState::Overview;
+				}
+				else {
+					self.view_data.clear();
+					result = result.state(State::List(false));
+				}
 			},
 		}
 		result.build()
@@ -96,12 +122,22 @@ impl<'s> ShowCommit<'s> {
 		let mut view_data = ViewData::new();
 		view_data.set_show_title(true);
 		view_data.set_show_help(true);
+		let view_builder_options = ViewBuilderOptions::new(
+			config.diff_tab_width as usize,
+			config.diff_tab_symbol.as_str(),
+			config.diff_space_symbol.as_str(),
+			config.diff_show_whitespace == DiffShowWhitespaceSetting::Both
+				|| config.diff_show_whitespace == DiffShowWhitespaceSetting::Leading,
+			config.diff_show_whitespace == DiffShowWhitespaceSetting::Both
+				|| config.diff_show_whitespace == DiffShowWhitespaceSetting::Trailing,
+		);
 		Self {
 			commit: None,
 			config,
 			no_commit_view_data: ViewData::new_error("Not commit data to show"),
+			state: ShowCommitState::Overview,
+			view_builder: ViewBuilder::new(view_builder_options, &config.key_bindings),
 			view_data,
-			view_builder: ViewBuilder::new(&config.key_bindings),
 		}
 	}
 
@@ -132,8 +168,16 @@ impl<'s> ShowCommit<'s> {
 						),
 					]));
 
-					self.view_builder
-						.build_view_data_for_overview(&mut self.view_data, commit, is_full_width);
+					match self.state {
+						ShowCommitState::Overview => {
+							self.view_builder
+								.build_view_data_for_overview(&mut self.view_data, commit, is_full_width);
+						},
+						ShowCommitState::Diff => {
+							self.view_builder
+								.build_view_data_diff(&mut self.view_data, &commit, is_full_width)
+						},
+					}
 					self.view_data.set_view_size(view_width, view_height);
 					self.view_data.rebuild();
 				}